requirements.txt
aiohttp 
aiosmtplib 
aiomysql 
boto3
python-dotenv
cryptography